Refactored helper functions into traits, updated php doc dcomments
