+++ src/Service/Studio/ConfigurationService.php
@@ -13,19 +13,19 @@ declare(strict_types=1);
 
 namespace Pimcore\Bundle\DataImporterBundle\Service\Studio;
 
-use Pimcore\Bundle\DataHubBundle\Configuration;
 use Pimcore\Bundle\DataImporterBundle\Hydrator\ConfigurationDetailHydratorInterface;
 use Pimcore\Bundle\DataImporterBundle\Schema\ConfigurationDetail;
+use Pimcore\Bundle\DataImporterBundle\Service\Studio\Traits\ConfigurationPermissionTrait;
 use Pimcore\Bundle\DataImporterBundle\Utils\Constants\PermissionConstants;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ConflictException;
-use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ForbiddenException;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * @internal
  */
 final readonly class ConfigurationService implements ConfigurationServiceInterface
 {
+    use ConfigurationPermissionTrait;
+
     public function __construct(
         private ConfigurationDetailHydratorInterface $configurationDetailHydrator
     ) {
@@ -64,30 +64,5 @@ final readonly class ConfigurationService implements ConfigurationServiceInterfa
         $config->save();
 
         return $this->configurationDetailHydrator->hydrate($config);
-    }
-
-    /**
-     * Load and validate a configuration by name, checking the given permission.
-     *
-     * @throws NotFoundHttpException if the configuration does not exist
-     * @throws ForbiddenException if the current user lacks the required permission
-     */
-    private function loadConfigurationWithPermission(string $name, string $permission): Configuration
-    {
-        $config = Configuration::getByName($name);
-
-        if (!$config) {
-            throw new NotFoundHttpException(
-                sprintf('Configuration with name "%s" not found', $name)
-            );
-        }
-
-        if (!$config->isAllowed($permission)) {
-            throw new ForbiddenException(
-                sprintf('Access denied to configuration "%s"', $name)
-            );
-        }
-
-        return $config;
     }
 }
