+++ src/Service/Studio/TransformationService.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Pimcore\Bundle\DataImporterBundle\Service\Studio;
 
-use Pimcore\Bundle\DataHubBundle\Configuration;
 use Pimcore\Bundle\DataImporterBundle\DataSource\Interpreter\InterpreterFactory;
 use Pimcore\Bundle\DataImporterBundle\Event\Studio\PreResponse\TransformationResultPreviewsEvent;
 use Pimcore\Bundle\DataImporterBundle\Event\Studio\PreResponse\TransformationResultTypeEvent;
@@ -23,20 +22,21 @@ use Pimcore\Bundle\DataImporterBundle\Preview\PreviewService;
 use Pimcore\Bundle\DataImporterBundle\Processing\ImportProcessingService;
 use Pimcore\Bundle\DataImporterBundle\Schema\TransformationResultPreviewsResponse;
 use Pimcore\Bundle\DataImporterBundle\Schema\TransformationResultTypeResponse;
+use Pimcore\Bundle\DataImporterBundle\Service\Studio\Traits\ConfigurationPermissionTrait;
+use Pimcore\Bundle\DataImporterBundle\Service\Studio\Traits\CurrentUserResolverTrait;
 use Pimcore\Bundle\DataImporterBundle\Settings\ConfigurationPreparationService;
 use Pimcore\Bundle\DataImporterBundle\Utils\Constants\PermissionConstants;
-use Pimcore\Bundle\StudioBackendBundle\Exception\Api\EnvironmentException;
-use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ForbiddenException;
 use Pimcore\Bundle\StudioBackendBundle\Security\Service\SecurityServiceInterface;
-use Pimcore\Model\User;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * @internal
  */
 final readonly class TransformationService implements TransformationServiceInterface
 {
+    use ConfigurationPermissionTrait;
+    use CurrentUserResolverTrait;
+
     public function __construct(
         private TransformationHydratorInterface $transformationHydrator,
         private PreviewService $previewService,
@@ -132,46 +132,5 @@ final readonly class TransformationService implements TransformationServiceInter
         );
 
         return $response;
-    }
-
-    /**
-     * Load and validate a configuration by name, checking the given permission.
-     *
-     * @throws NotFoundHttpException if the configuration does not exist
-     * @throws ForbiddenException if the current user lacks the required permission
-     */
-    private function loadConfigurationWithPermission(string $name, string $permission): Configuration
-    {
-        $config = Configuration::getByName($name);
-
-        if (!$config) {
-            throw new NotFoundHttpException(
-                sprintf('Configuration with name "%s" not found', $name)
-            );
-        }
-
-        if (!$config->isAllowed($permission)) {
-            throw new ForbiddenException(
-                sprintf('Access denied to configuration "%s"', $name)
-            );
-        }
-
-        return $config;
-    }
-
-    /**
-     * Resolve the current user, ensuring it is a Pimcore User instance.
-     *
-     * @throws EnvironmentException if the current user cannot be resolved
-     */
-    private function resolveCurrentUser(): User
-    {
-        $user = $this->securityService->getCurrentUser();
-
-        if (!$user instanceof User) {
-            throw new EnvironmentException('Could not resolve current user');
-        }
-
-        return $user;
     }
 }
